+++ src/project.py
@@ -23,6 +23,7 @@ class Project:
         # self.wip_path = None
         # self.board_top_path = None
         # self.board_bot_path = None
+        self.refproj_path = ""
         self.loading = False
 
     def to_serializable(self) -> dict:
@@ -34,6 +35,7 @@ class Project:
             'pnp_separator': self.pnp_separator,
             'pnp_first_row': self.pnp_first_row,
             'pnp_columns': self.pnp_columns.serialize(),
+            'refproj_path': self.refproj_path,
             'pnp_grid': grid,
         }
         return ret
@@ -46,6 +48,7 @@ class Project:
             self.pnp_separator = inp['pnp_separator']
             self.pnp_first_row = inp['pnp_first_row']
             self.pnp_columns.deserialize(inp['pnp_columns'])
+            self.refproj_path = inp.get('refproj_path', '')
             self.pnp_grid = text_grid.TextGrid()
             self.pnp_grid.from_serializable(inp['pnp_grid'])
         except Exception as e: